// ui/src/api/gen/DiffSpec.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DiffSpec = { "kind": "commit", id: string, parent: number, } | { "kind": "compare", from: string, to: string, } | { "kind": "worktree", from: string, worktree: string, } | { "kind": "wip", worktree: string, staged: boolean, };