// ui/src/api/gen/FileChange.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BlobSource } from "./BlobSource";

export type FileChange = { path: string, 
/**
 * The source path of a rename or copy.
 */
oldPath: string | null, 
/**
 * `A`, `C`, `D`, `M`, `R`, `T`, `U` or `X`.
 */
status: string, 
/**
 * `None` for binary files.
 */
additions: number | null, deletions: number | null, old: BlobSource, new: BlobSource, submodule: boolean, };